Refactor(gameRestartController): gameController의 기능을 쪼개서 분담
